move __class__ into parent class
