+++ spartan/model/_model.py
@@ -22,7 +22,7 @@ class Model():
         self.model_name = model_name
 
     @classmethod
-    def __create__(self, data: object, params: dict):
+    def __create__(cls, data: object, params: dict):
         '''Interface of creation by class, overrided by subclasses.
 
         Raises
@@ -36,6 +36,14 @@ class Model():
 class DMmodel(Model):
     '''Interface for data mining model.
     '''
+
+    def __init__(self, tensor, params):
+        pass
+
+    @classmethod
+    def __create__(cls, tensor, params):
+        _obj = cls(tensor, params)
+        return _obj
 
     def run(self):
         '''Interface of run function, overrided by subclasses.
@@ -51,6 +59,14 @@ class DMmodel(Model):
 class MLmodel(Model):
     '''Interface for machine learning model.
     '''
+
+    def __init__(self, tensor, params):
+        pass
+
+    @classmethod
+    def __create__(cls, tensor, params):
+        _obj = cls(tensor, params)
+        return _obj
 
     def fit(self):
         '''Interface of fit function, overrided by subclasses.
