+++ spartan/model/beatlex/__init__.py
@@ -1,0 +1,4 @@
+from .Beatlex import BeatLex
+
+def __call__():
+    return BeatLex
