+++ spartan/task/summarization.py
@@ -33,4 +33,4 @@ class Summarization(Task):
 class SumPolicy(Enum):
     '''Registration for path of models who can do summarization task.
     '''
-    Beatlex = MODEL_PATH + ".beatlex.Beatlex"
+    Beatlex = MODEL_PATH + ".beatlex"
